pyproject.toml
[project]
name = "mindmap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "markdown-it-py",
    "pyyaml",
]
